+++ fastai/imports/core.py
@@ -3,7 +3,7 @@ import math, matplotlib.pyplot as plt, numpy as np, pandas as pd, random
 import scipy.stats, scipy.special
 import abc, collections, hashlib, itertools, json, operator, pathlib
 import mimetypes, inspect, typing, functools, importlib, weakref
-import html, re, requests, tarfile, numbers, tempfile
+import html, re, requests, tarfile, numbers, tempfile, bz2
 
 from abc import abstractmethod, abstractproperty
 from collections import abc,  Counter, defaultdict, Iterable, namedtuple, OrderedDict